+++ src/main/java/com/codeit/monew/domain/article/dto/response/ArticleDto.java
@@ -1,4 +1,4 @@
-package com.codeit.monew.domain.article.dto;
+package com.codeit.monew.domain.article.dto.response;
 
 import com.codeit.monew.domain.article.ArticleSource;
 import io.swagger.v3.oas.annotations.media.Schema;
